change app structure
