+++ src/main/java/app/Main.java
@@ -1,0 +1,13 @@
+package app;
+
+import app.index.Indexer;
+
+import java.util.HashMap;
+import java.util.Map;
+
+public class Main {
+
+    public static void main(String[] args) {
+
+    }
+}
