+++ src/main/java/app/search/Searcher.java
@@ -1,2 +1,4 @@
+package app.search;
+
 public class Searcher {
 }
